Change value of valid_to column
The value of valid_to column was set to null to denote infinity
but it would make searching data a bit cumbersome.

e.g. when AS OF clause will be implemented, the right thing to do
is equate if valid_to is greater than entered date but the value
is either null or somewhere in the past.

Using '9999-12-31' would denote infinity and make implementation
of search clauses more easy.