Fix bug where deletion query updated all rows
1. Removed set_trace cuz sorry i forget to delete it

2. DeleteQueryBuilder built an update query which only had one
condition in its WHERE clause, the primary key. This resulted in
the query updating the previous rows with that id to have the same
value for valid_to column. Fixed it by adding a second condition
that checks the time is infinity to only update the latest record
and not the previous ones.