Refactor project
Changes:

Updated imports from being relative to absolute.

SelectQueryHandler now uses regex to parse and figure out where
to pass the flow.

Empty out QueryParser functions as they'll be implemented to use
regex. I've modified one function to use regex.

QueryParser functions and SelectQueryHandler were the only
functions using QueryParser.query as a tuple but since they have
been updated to use regex, there is no need for QueryParser.query
to be a tuple so now QueryParser.query returns a string instead of
a tuple. This removes the extra step of converting the query to a
string before being parsed.