Add temporal query tests
Temporal query tests weren't added before because the local time
variable was being created inside the query builder class, so the
test would fail because the local time variable in the test and
the local time variable in the class were different.

To tackle this, I've modified the query builder classes so they
now accept a new argument which is local time. This way the query
that gets created uses the passed time string instead of creating
its own, thereby allowing us to have tests for temporal queries.