Add create_column_value_dictionary function
It accepts a list with column and values with odd index as column
and even index as value. It creates a dictionary with keys as
column name and key value as column value.

It is part of the support being added to allow multiple column
update in single update query.